requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=0.19.0
Jinja2>=3.0.0
pydantic-settings